% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{array_to_HTMLTable}
\alias{array_to_HTMLTable}
\title{Render a list as an HTML table
Useful to display data in an Rmarkdown exercice}
\usage{
array_to_HTMLTable(c, columns)
}
\description{
Render a list as an HTML table
Useful to display data in an Rmarkdown exercice
}
