% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{array_to_MDTable}
\alias{array_to_MDTable}
\title{Render a list as an markdown table
Useful to display data in an Rmarkdown exercice}
\usage{
array_to_MDTable(c, columns)
}
\description{
Render a list as an markdown table
Useful to display data in an Rmarkdown exercice
}
