% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generation.R
\name{backup_config}
\alias{backup_config}
\title{Make a backup copy of the current configuration file}
\usage{
backup_config()
}
\description{
This is particularly useful after generating the exam sheets
because some output files need to be re-used in the
optical correction step.
}
