% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correction.R
\name{check_evaluation}
\alias{check_evaluation}
\title{Check some necessary conditions for the automatic
evaluation of the scanned exams}
\usage{
check_evaluation()
}
\description{
Check some necessary conditions for the automatic
evaluation of the scanned exams
}
