% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{check_solution}
\alias{check_solution}
\title{Check that the automatic solution generation function is well written.
It is useful when writing exercices}
\usage{
check_solution(gen_sol, params, solution)
}
\arguments{
\item{gen_sol}{function that generates the solution}

\item{params}{named list of arguments of gen_sol}

\item{solution}{the expected solution that should be calculated when calling gen_sol with the parameters params}
}
\description{
Check that the automatic solution generation function is well written.
It is useful when writing exercices
}
