% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{count_decimals}
\alias{count_decimals}
\title{Return the length of the (significant) decimal part of x
It can be useful when writing exercices}
\usage{
count_decimals(x)
}
\description{
Return the length of the (significant) decimal part of x
It can be useful when writing exercices
}
