% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{create_config_dir}
\alias{create_config_dir}
\title{Create a configuration dir in the current folder
The configuration dir contains the LateX templates for the handwritten exams.
If the current working directory contains a Configuration folder its LateX templates are used.
This allows the modification of the templates}
\usage{
create_config_dir()
}
\description{
Create a configuration dir in the current folder
The configuration dir contains the LateX templates for the handwritten exams.
If the current working directory contains a Configuration folder its LateX templates are used.
This allows the modification of the templates
}
