% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correction.R
\name{evaluation}
\alias{evaluation}
\title{call the evaluation function of rexams
and generate the results.}
\usage{
evaluation()
}
\description{
call the evaluation function of rexams
and generate the results.
}
