% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correction.R
\name{evaluation_summary}
\alias{evaluation_summary}
\title{Generates a minimal result table.
It contains only 3 columns StudentName, ID, Note}
\usage{
evaluation_summary()
}
\description{
Generates a minimal result table.
It contains only 3 columns StudentName, ID, Note
}
