% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{load_exam_config}
\alias{load_exam_config}
\title{Load the exam config file.
This contains all the information needed for the exam generation.
This function should be called at the beginning of every generation and correction script}
\usage{
load_exam_config(config_file)
}
\arguments{
\item{config_file}{The path to the config file}
}
\description{
Load the exam config file.
This contains all the information needed for the exam generation.
This function should be called at the beginning of every generation and correction script
}
