% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{make_exam}
\alias{make_exam}
\title{Generate NOPS (exams sheets)
Before using, load a configuration file with load_exam_config.}
\usage{
make_exam()
}
\description{
Generate NOPS (exams sheets)
Before using, load a configuration file with load_exam_config.
}
