% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generation.R
\name{make_traditional_exams}
\alias{make_traditional_exams}
\title{Generate traditional exam sheets (no optical correction)}
\usage{
make_traditional_exams()
}
\description{
Generate traditional exam sheets (no optical correction)
}
