% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{package}
\name{morexams}
\alias{morexams}
\title{morexams: an easy start with rexam}
\description{
rexam (http://www.r-exams.org/) is a very powerful R package for exam sheets geenration and correction.
The package is reach and its functions are highly customizable.
}
\details{
For this very reason its use might be confusing at first. This package is the synthesis of one year wokring with r-exams. `morexams` is built to facilitate users access to the `rexam` functionality, possibly lowering the entry cost.
`morexams` wraps `rexam` and adds mainly three features:

* add checks to avoid simple problems (duplicate students, missing sheets ...) which are difficult to debug in `rexam`

* add an `init` function to initialize a working folder with a basic structure needed to use `rexams` out of the box. The `init` function generates a fully-working example to immediately start with `rexam`.

* move the configuration from the function parameters to a configuration file, useful to store configurations and simplyfy the script writing.
}
