% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{morexams_init}
\alias{morexams_init}
\title{Initialize the current folder with the default working three of morexams
Tutorial files are also copied in the working folder.}
\usage{
morexams_init()
}
\description{
Initialize the current folder with the default working three of morexams
Tutorial files are also copied in the working folder.
}
\examples{
library(morexams)
morexams_init()
}
