% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correction.R
\name{read_nops_scans}
\alias{read_nops_scans}
\title{Read the multiple choice part of the scanned NOPS
The scans should be PNG images
They must be placed in the "Scan/nops" folder}
\usage{
read_nops_scans()
}
\description{
Read the multiple choice part of the scanned NOPS
The scans should be PNG images
They must be placed in the "Scan/nops" folder
}
