% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correction.R
\name{read_string_scan}
\alias{read_string_scan}
\title{Read the written part of the scanned NOPS
The scans should be PNG images
image files must be placed in the "Scan/string" folder}
\usage{
read_string_scan()
}
\description{
Read the written part of the scanned NOPS
The scans should be PNG images
image files must be placed in the "Scan/string" folder
}
