% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run_evaluation}
\alias{run_evaluation}
\title{Run the evaluation of the scanned and read NOPS
if force_read_scan is TRUE the scans are reading each time
the funcion is called (time consuming), otherwise it uses the last reading}
\usage{
run_evaluation(force_read_scan = FALSE)
}
\arguments{
\item{force_read_scan}{if FALSE, the image processing takes place only the first time this function is called, further calls use the results of the first image processing}
}
\description{
Run the evaluation of the scanned and read NOPS
if force_read_scan is TRUE the scans are reading each time
the funcion is called (time consuming), otherwise it uses the last reading
}
