% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{safe_mkdir}
\alias{safe_mkdir}
\title{Remove the specified folder.}
\usage{
safe_mkdir(folder)
}
\description{
Check if the folder exists already.
If it exists, it will continue according to the
value of the OVERWRITE config variable
}
