% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generation.R
\name{write_one_solution}
\alias{write_one_solution}
\title{Write the solution of the exam in a PDF document
The NOPS need to to have already been generated.}
\usage{
write_one_solution()
}
\description{
Write the solution of the exam in a PDF document
The NOPS need to to have already been generated.
}
